refactor: improve styles
