+++ src/app/detailed-card/_components/DetailedCardContent/SimilarProducts/index.tsx
@@ -1,5 +1,5 @@
 "use client";
-import ProductCardCarouselSecondary from "@/_components/ProductCardCarousel/ProductCardCarouselSecondary";
+import Index from "@/_components/ProductCardCarousel";
 import productCardsData from "@/_data/mockProductCardsData/productCardData.json";
 
 const SimilarProducts = () => {
@@ -9,7 +9,7 @@ const SimilarProducts = () => {
         <h2 className="font-bold text-3xl text-center md:text-start md:text-mm-title-two mb-9">
           Похожие товары
         </h2>
-        <ProductCardCarouselSecondary cardsContent={productCardsData} />
+        <Index cardsContent={productCardsData} />
       </div>
     </section>
   );
